refactor BookKeeper in order to make it mockable add testscases for builders

+++ bookkeeper-server/src/main/java/org/apache/bookkeeper/client/api/CreateAdvBuilder.java
@@ -25,7 +25,7 @@ import java.util.concurrent.CompletableFuture;
  */
 public interface CreateAdvBuilder extends OpBuilder<WriteAdvHandle> {
 
-    CreateAdvBuilder withLedgerId(int ledgerId);
+    CreateAdvBuilder withLedgerId(long ledgerId);
 
     @Override
     CompletableFuture<WriteAdvHandle> execute();
